+++ src/app/components/transfers/live-stats-table.js
@@ -33,7 +33,7 @@ const COMMON_I18N_PREFIX = 'components.transfers.';
 const I18N_PREFIX = COMMON_I18N_PREFIX + 'liveTableStats.';
 
 const tableExcludedColumns = {
-  scheduled: ['startedAt', 'finishedAt', 'totalBytes', 'totalFiles', 'status'],
+  scheduled: ['startedAt', 'finishedAt', 'totalBytes', 'totalFiles'],
   current: ['scheduledAt', 'finishedAt'],
   completed: ['scheduledAt'],
 };
@@ -46,7 +46,7 @@ export default Component.extend({
 
   /**
    * @virtual 
-   * @type {Array<Transfer>}
+   * @type {ReplacingChunksArray<Transfer>}
    */
   transfers: undefined,
 
@@ -88,12 +88,28 @@ export default Component.extend({
    * @type {Array<string>|undefined} array of transfer ids
    */
   selectedTransferIds: undefined,
-      
+    
+  /**
+   * @virtual
+   * If true, table was just showed to user, so some additional
+   * styles/operations, etc. should be performed
+   * @type {boolean}
+   */
+  justOpened: false,
+  
+  /**
+   * @virtual
+   * Invoke this function after operations associated with fresh opened
+   * table have been finished (see `justOpened` flag)
+   * @type {Function}
+   */
+  clearJustOpened: () => {},
+  
   /**
    * @type {EmberObject}
    */
   firstRowSpace: undefined,
-  
+    
   /**
    * If true, component is rendered in mobile mode.
    * @type {boolean}
@@ -112,6 +128,14 @@ export default Component.extend({
   _stickyTableHeaderOffset: 0,
   
   /**
+   * True if the list is reloaded.
+   * Note, it is still false if fetching more data or particular records 
+   * are reloaded.
+   * @type {Ember.ComputedProperty<boolean>}
+   */
+  _isReloading: computed.reads('transfers.isReloading'),
+  
+  /**
    * Custom icons for ember-models-table addon.
    * @type {Ember.Object}
    */
@@ -120,6 +144,12 @@ export default Component.extend({
     'sort-desc': 'oneicon oneicon-arrow-down',
   }),
 
+  /**
+   * True if table has no data (except virtual rows)
+   * @type {Ember.ComputedProperty<boolean>}
+   */
+  _tableDataEmpty: computed.lt('_tableData.length', 2),
+  
   /**
    * Custom classes for ember-models-table addon.
    * @type {Ember.Object}
@@ -162,6 +192,7 @@ export default Component.extend({
         firstRowSpace,
         movedTransfers,
         updaterId,
+        transferType,
       } = this.getProperties(
         'transfers',
         'providers',
@@ -170,7 +201,8 @@ export default Component.extend({
         '_tableDataCache',
         'firstRowSpace',
         'movedTransfers',
-        'updaterId'
+        'updaterId',
+        'transferType'
       );
       
       if (transfers && providers) {
@@ -210,6 +242,7 @@ export default Component.extend({
               providers,
               providersColors,
               updaterId,
+              transferCollection: transferType,
             }));
             arrayChanged = true;
           }
@@ -334,7 +367,12 @@ export default Component.extend({
       'notifyTransferListChanged',
       this.get('transferType')
     );
-
+  }),
+  
+  isReloadingFinished: observer('_isReloading', function () {
+    if (this.get('justOpened') && !this.get('_isReloading')) {
+      this.get('clearJustOpened')();
+    }
   }),
   
   init() {
@@ -357,6 +395,8 @@ export default Component.extend({
     
     _resizeEventHandler();
     _window.addEventListener('resize', _resizeEventHandler);
+    
+    this.notifyTransferListChanged();
   },
   
   willDestroyElement() {