+++ src/app/utils/fake-list-record-relation.js
@@ -14,16 +14,30 @@ import Ember from 'ember';
 const {
   computed,
   get,
+  assert,
 } = Ember;
 
 export default PromiseObject.extend({
+  /**
+   * Initialized in init
+   * @type {FakeListRecord}
+   */
+  _fakeListRecord: undefined,
+  
   isLoading: computed.reads('isPending'),
   isLoaded: computed.not('isLoading'),
+    
   reload() {
     return this.get('_fakeListRecord').reload(...arguments);
   },
+  
   init() {
     this._super(...arguments);
+    const initChunksArray = this.get('initChunksArray');
+    assert(
+      'utils:fake-list-record-relation: initChunksArray is required',
+      initChunksArray != null
+    );
     const _fakeListRecord = this.set('_fakeListRecord', FakeListRecord.create({
       initChunksArray: this.get('initChunksArray'),
     }));